restructure User
