+++ model/user.go
@@ -1,7 +1,6 @@
 package model
 
 import (
-	"github.com/jmoiron/sqlx"
 	"time"
 )
 
@@ -14,30 +13,10 @@ type User struct {
 	DeletedAt *time.Time `db:"deleted_at"`
 }
 
-// UserModel -- have method about user table in database
-type UserModel struct {
-	db *sqlx.DB
-}
-
 // UserModelImpl -- Define database control methods about user
 type UserModelImpl interface {
 	All() ([]User, error)
 	// Create(user *User) error
 	// Delete(user *User) error
 	// FindByID(id uint) (User, error)
-}
-
-// NewUserModel UserModelの初期化
-func NewUserModel(db *sqlx.DB) *UserModel {
-	return &UserModel{db: db}
-}
-
-// All -- SELECT * FROM users
-func (u *UserModel) All() ([]User, error) {
-	users := []User{}
-	err := u.db.Select(&users, "SELECT * from user")
-	if err != nil {
-		return nil, err
-	}
-	return users, nil
 }
