+++ route/router.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"github.com/Amakuchisan/QuestionBox/database"
 	"github.com/Amakuchisan/QuestionBox/handler"
-	"github.com/Amakuchisan/QuestionBox/model"
+	"github.com/Amakuchisan/QuestionBox/repository"
 	"github.com/labstack/echo"
 	"html/template"
 	"io"
@@ -41,7 +41,7 @@ func Init() *echo.Echo {
 	e.GET("/auth/login/:provider", handler.LoginHandler)
 	e.GET("/auth/callback/:provider", handler.CallbackHandler)
 
-	userHandler := handler.NewUserHandler(model.NewUserModel(database.DB))
+	userHandler := handler.NewUserHandler(repository.NewUserModel(database.DB))
 	e.GET("/users", userHandler.UserAll)
 	// e.POST("/users", userHandler.CreateUser)
 	// e.GET("/users/:id", userHandler.DetailUser)
